app/install/requirements.txt
python-dotenv
flask
flask-sqlalchemy
flask-bcrypt
flask-cors
flask-session
redis
PyJWT
pandas
langchain_huggingface
langchain_core
langchain_weaviate
weaviate-client
pymongo
selenium
